requirements.txt
sqlalchemy==1.4.22
myst-nb~=0.13.1
sphinx-book-theme
sphinx-design
sphinx-copybutton
sphinx-thebe
